Create specs defining the expected state of cells in a grid
Defining the expectations for the perspective of `cell_0_0` generated
the expected results. Defining the expectations for the perspective of
`cell_1_0` highlighted problems with how relationships between cells
were managed. The `getRelatedLocations` function was intended to enable
updating the adjacent cells as the relationships were defined. Though
this did not work as intended which is why the loop using the data from
`getRelatedLocations` is commented out.