Create the `Game` structure
The `Game` generates the list of `Cell` instances. The list of `moves`
was intended to reflect that the state of the game can be derived from a
list of recorded moves based on the state of cells.

The mutations on `Cell` were a concession to time because this
information is calculable from the `Cell`s and moves as the list of
`Cell` is not intended to be modified once it is created.