Create `NOTES.md` outlining intentions rather than results
The notes file was created in the last 15-20 minutes of the challenge in
an attempt to outline the direction taken in the challenge. The contents
of `NOTES.md` would have been useful as a place to start rather than
starting with data modeling as I did. `NOTES.md` outlines the intended
interactions with the game.