Create a version of minesweeper using mogwai
Uses "game" components to display minesweeper (like) game board.  The
"game" is based on the API provided by `@mines/nest` which defines
endpoints for listing games, viewing a specific game's state and
executing a "move" in the game. The game listing screen usings the API
to fetch the list of game ids and display them.

The `route_dispatch` module functions can be used to turn a
`Transmitter` into a `ViewBuilder` or to push a new entry into browser
history. The `Transmitter` is needed to enable navigation to a new
`Route` from the `GameList`.

Changes to the game are sent through the `api::patch_game` function
which sends a PATCH to the server, the server responds with an updated
board state. The board view is replaced on each update.