Use `wire_filter_fold` to transmit only when status changes
Use a `branch_map` from the `Receiver` that only updates when game
status changes in order to generate a view patch for each game status
change. This was previously being attempted with a `wire_filter_fold`
but the initial value was not being mutated which is what
`wire_filter_fold` expects.