Only create the board once per game
Send the whole new list of cells into the board each time via
`Receiver<CellUpdate>` but only re-create the board when the `GameId`
changes. This relies on the `board_cell` function or the `BoardCell`
component to only trigger updates for the cell when the cell requires a
change. As implemented, this is still inefficient because it updates
every `BoardCell` on every new `GameState` but the whole `<table>` isn't
dropped and recreated.