Reorganize where `.env` lives to move root `Cargo.toml`
In order to move `Cargo.toml` to the project root the `.env` file needs
to be readable from the root directory so that the `dotenv_codegen`
crate can find it (it appears to looks from the `Cargo.toml` directory)
but the `@mines/nest` package needs it to be in the `server-nest`
directory. In order to accomplish both of these objectives use a single
`.env` file to hold all the environment variables pulled from SSM and
copy the generated `.env` file into the other locations via the
`Makefile`.